+++ dev-setup.sh
@@ -2,12 +2,10 @@
 
 # clean up current situation
 docker-compose stop
-docker-compose rm -v --force
+docker-compose rm -v --force --all
 find . \( -name "*.pyc" -o -name "*.pyo" \) -print0 | xargs -0 rm -rf
 
 # rebuild containers
-docker-compose stop
-docker-compose rm -f -v
 docker-compose build
 
 # start data containers